Extract shell host projection services
