+++ clients/apple/scripts/test-shell-runtime-metadata.sh
@@ -16,6 +16,8 @@ CLANG_MODULE_CACHE_PATH="$MODULE_CACHE_DIR" swiftc \
     "$REPO_ROOT/clients/apple/AlanNative/Models/Shell/ShellStateMutations.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/ShellModel.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/ShellControlPlane.swift" \
+    "$REPO_ROOT/clients/apple/AlanNative/Services/Shell/ShellPaneProjectionService.swift" \
+    "$REPO_ROOT/clients/apple/AlanNative/Services/Shell/ShellStatePersistenceStore.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/TerminalHostRuntime.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/TerminalSurfaceController.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/TerminalRuntimeService.swift" \
